requirements.txt
openai
chromadb
sentence-transformers
tqdm
python-dotenv
sqlight-utils
feedparser
newspaper3k